returning response entities, adjusted tests, finished project

+++ 5JavaSpring/demo/src/main/java/com/example/demo/classEmployee/ClassEmployee.java
@@ -72,7 +72,7 @@ public class ClassEmployee {
             return 0.0;
         }
 
-        return employees.size() / maxEmployeeListSize * 100;
+        return (double) employees.size() / (double) maxEmployeeListSize * 100;
     }
